+++ DependencyInjection/Configuration.php
@@ -98,6 +98,10 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('uri_override')
                     ->cannotBeEmpty()
                 ->end()
+                ->scalarNode('http_client')
+                  ->info('Override the default HTTP client with a custom Guzzle instance. Service ID as string.')
+                  ->cannotBeEmpty()
+                ->end()
             ->end()
         ;