+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Exposed new features from the newer Contentful SDK
   * Overriding URI used to communicate with the Contentful API (`uri_override`)
   * Setting the default locale for the client (`default_locale`)
+  * Setting a custom Guzzle instance (`http_client`)
 
 ### Changed
 * Use version 0.7 of the Contentful SDK. **Note:** This release contains breaking changes, you can read more about them